+++ SaSH/form/petinfoform.h
@@ -26,7 +26,7 @@ class PetInfoForm : public QWidget, public Indexer
 	Q_OBJECT
 
 public:
-	explicit PetInfoForm(qint64 index, QWidget* parent);
+	explicit PetInfoForm(long long index, QWidget* parent);
 
 	virtual ~PetInfoForm();
 
@@ -49,15 +49,15 @@ protected:
 private:
 	Ui::PetInfoFormClass ui;
 
-	qint64 base_level_ = 0;
-	qint64 base_maxHp_ = 0;
-	qint64 base_atk_ = 0;
-	qint64 base_def_ = 0;
-	qint64 base_agi_ = 0;
+	long long base_level_ = 0;
+	long long base_maxHp_ = 0;
+	long long base_atk_ = 0;
+	long long base_def_ = 0;
+	long long base_agi_ = 0;
 
-	qint64 cur_level_ = 0;
-	qint64 cur_maxHp_ = 0;
-	qint64 cur_atk_ = 0;
-	qint64 cur_def_ = 0;
-	qint64 cur_agi_ = 0;
+	long long cur_level_ = 0;
+	long long cur_maxHp_ = 0;
+	long long cur_atk_ = 0;
+	long long cur_def_ = 0;
+	long long cur_agi_ = 0;
 };
