+++ SaSH/model/fastlabel.h
@@ -68,7 +68,7 @@ private:
 	QString content_msg_;
 	QColor text_color_ = Qt::black;
 	QColor background_color_ = QColor("#F1F1F1");
-	qint64 new_font_size_ = 12;
+	long long new_font_size_ = 12;
 };
 #endif // MYLABEL_H
