+++ SaSH/model/indexer.h
@@ -24,14 +24,14 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 class Indexer
 {
 public:
-	explicit Indexer(qint64 index) : index_(index) {}
+	explicit Indexer(long long index) : index_(index) {}
 
-	virtual inline void setIndex(qint64 index)
+	virtual inline void setIndex(long long index)
 	{
 		index_.store(index, std::memory_order_release);
 	}
 
-	virtual inline qint64 getIndex() const { return index_.load(std::memory_order_acquire); }
+	virtual inline long long getIndex() const { return index_.load(std::memory_order_acquire); }
 
 private:
 	std::atomic_int64_t index_ = -1;
